rename structs to iterator_trait
