+++ src/iterator_trait.rs
