+++ src/lib.rs
@@ -4,11 +4,11 @@ mod db;
 mod entry;
 mod error;
 mod format;
+mod iterator_trait;
 mod levels;
 mod memtable;
 mod ops;
 mod opt;
-mod structs;
 mod table;
 mod util;
 mod value;
@@ -22,5 +22,5 @@ pub use value::Value;
 
 pub use db::{Agate, AgateOptions};
 pub use error::{Error, Result};
+pub use iterator_trait::AgateIterator;
 pub use skiplist::Skiplist;
-pub use structs::AgateIterator;
